Refactor(Security, Api): 코드 리팩토링
- 안쓰는 클래스 삭제
- System.out.println -> log.info로 교체
- 클래스명 변경
